Restructure project and README.

+++ sources/get_next_line.c
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/14 01:52:11 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2021/02/22 00:42:40 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2021/03/26 02:06:16 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,8 +60,8 @@ static char	*read_and_join(int fd,
 							char *line_buffer,
 							bool *found_linebreak)
 {
-	char	*linebreak_position;
-	size_t	bytes_read;
+	char		*linebreak_position;
+	long int	bytes_read;
 
 	while (*found_linebreak == false)
 	{
